Trying to restructure the particle tracking code so that workflow has a simple structure:
crop:preprocess:track:link:postprocess:annotate

The idea is you will have a daughter class from this which you select options and it will then do in order. Sending this daughter to the TrackingGui will result in the visual output being displayed. You will also be able to process all or part.

The config settings will be stored in a dictionary as before.

I think there is a need to remove certain parts of ParticleTracker to a higher level such as the movie and datastore and preprocessor.